backend/requirements.txt
uvicorn
gunicorn
fastapi==0.99.1
tiktoken==0.5.1
openai==0.28.1
python-dotenv==1.0.0
python-multipart
python-socketio==5.10.0